+++ src/main/java/com/justdo/glue/sticker/domain/poststicker/service/PostStickerCommandService.java
@@ -3,5 +3,5 @@ package com.justdo.glue.sticker.domain.poststicker.service;
 import com.justdo.glue.sticker.domain.poststicker.dto.PostStickerDTO;
 
 public interface PostStickerCommandService {
-    PostStickerDTO.PostStickerItem BuildPostSticker(Long postId, Long stickerId, Double xLocation, Double yLocation, Double width, Double height, Double angle);
+    PostStickerDTO.PostStickerItem BuildPostSticker(Long postId, Long stickerId, int xLocation, int yLocation, int width, int height, int angle);
 }
